Split LLVM charset code out into llvm_charset.c
